Tests/LibWeb/Text/expected/abspos-relayout-move-static-position.txt

boundary=0,60,150,60
partial-layouts=0 full-layouts=1
FAIL
